完成用户管理功能调试；

+++ src/Dotnet9.WebAPI/Controllers/UserAdminController.cs
@@ -1,7 +1,6 @@
 ﻿using Dotnet9.EventBus;
 using Dotnet9.WebAPI.Application.Contracts.UserAdmin;
 using Dotnet9.WebAPI.Events;
-using Microsoft.EntityFrameworkCore;
 
 namespace Dotnet9.WebAPI.Controllers;
 
@@ -22,10 +21,17 @@ public class UserAdminController : ControllerBase
     }
 
     [HttpGet]
-    public Task<UserDTO[]> GetAllUsers()
+    public async Task<UserDTO[]> GetAllUsers()
     {
-        return _userManager.Users.Select(u => new UserDTO(u.Id, u.UserName, u.PhoneNumber, u.CreationTime))
-            .ToArrayAsync();
+        var userWithRoles = new List<UserDTO>();
+        foreach (var user in _userManager.Users.ToList())
+        {
+            var roleNames = await _userManager.GetRolesAsync(user);
+            userWithRoles.Add(new UserDTO(user.Id, user.UserName, roleNames.ToArray(), user.PhoneNumber,
+                user.CreationTime));
+        }
+
+        return userWithRoles.ToArray();
     }
 
     [HttpGet]
@@ -33,14 +39,14 @@ public class UserAdminController : ControllerBase
     public async Task<UserDTO> FindById(Guid id)
     {
         var user = await _userManager.FindByIdAsync(id.ToString());
-        return new UserDTO(user.Id, user.UserName, user.PhoneNumber, user.CreationTime);
+        var roleNames = await _userManager.GetRolesAsync(user);
+        return new UserDTO(user.Id, user.UserName, roleNames.ToArray(), user.PhoneNumber, user.CreationTime);
     }
 
     [HttpPost]
-    [Authorize(Roles = UserRoleConst.Admin)]
-    public async Task<ActionResult> AddAdminUser(AddAdminUserRequest req)
+    public async Task<ActionResult> AddUser(AddUserRequest req)
     {
-        var (result, user, password) = await _repository.AddAdminUserAsync(req.UserName, req.PhoneNumber);
+        var (result, user, password) = await _repository.AddUserAsync(req.UserName, req.RoleName, req.PhoneNumber);
         if (!result.Succeeded)
         {
             return BadRequest(result.Errors.SumErrors());
@@ -53,15 +59,21 @@ public class UserAdminController : ControllerBase
 
     [HttpDelete]
     [Route("{id}")]
-    public async Task<ActionResult> DeleteAdminUser(Guid id)
+    public async Task<ActionResult> DeleteUser(Guid id)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (userId == id.ToString())
+        {
+            return BadRequest("管理员不能删除自己");
+        }
+
         await _repository.RemoveUserAsync(id);
         return Ok();
     }
 
     [HttpPut]
     [Route("{id}")]
-    public async Task<ActionResult> UpdateAdminUser(Guid id, EditAdminUserRequest req)
+    public async Task<ActionResult> UpdateUser(Guid id, EditUserRequest req)
     {
         var user = await _repository.FindByIdAsync(id);
         if (user == null)
@@ -71,12 +83,22 @@ public class UserAdminController : ControllerBase
 
         user.PhoneNumber = req.PhoneNumber;
         await _userManager.UpdateAsync(user);
+        if (req.RoleName == UserRoleConst.User && await _userManager.IsInRoleAsync(user, UserRoleConst.Admin))
+        {
+            await _userManager.RemoveFromRoleAsync(user, UserRoleConst.Admin);
+        }
+        else if (req.RoleName == UserRoleConst.Admin &&
+                 await _userManager.IsInRoleAsync(user, UserRoleConst.Admin) == false)
+        {
+            await _userManager.AddToRoleAsync(user, UserRoleConst.Admin);
+        }
+
         return Ok();
     }
 
     [HttpPost]
     [Route("{id}")]
-    public async Task<ActionResult> ResetAdminUserPassword(Guid id)
+    public async Task<ActionResult> ResetUserPassword(Guid id)
     {
         var (result, user, password) = await _repository.ResetPasswordAsync(id);
         if (!result.Succeeded)
